+++ src/models/modifiers.ts
@@ -1,0 +1,6 @@
+/**
+ * The structure of the `modifiers.json` file.
+ */
+export interface Modifiers {
+  [category: string]: string[];
+}